make prod template default, rename default rapid-prototype
