--- templates/Production template/package.json
@@ -1,37 +1,0 @@
-{
-    "private": true,
-    "name": "lit-starter-kit",
-    "version": "1.0.0",
-    "main": "index.html",
-    "scripts": {
-        "dev": "vite",
-        "build": "vite build"
-    },
-    "dependencies": {
-        "@conectate/ct-lit": "^3.2.1",
-        "@conectate/ct-router": "^3.2.1",
-        "@webcomponents/webcomponentsjs": "^2.5.0",
-        "lit": "^2.0.0-rc.2"
-    },
-    "devDependencies": {
-        "@babel/core": "^7.14.3",
-        "@babel/plugin-proposal-decorators": "^7.14.2",
-        "@babel/plugin-proposal-export-default-from": "^7.12.13",
-        "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-        "@babel/plugin-syntax-import-meta": "^7.10.4",
-        "@babel/preset-env": "^7.14.4",
-        "@firebase/app-types": "^0.6.2",
-        "@rollup/plugin-babel": "^5.3.0",
-        "@rollup/plugin-node-resolve": "^13.0.0",
-        "babel-plugin-bundled-import-meta": "^0.3.2",
-        "babel-plugin-template-html-minifier": "^4.1.0",
-        "prettier": "^2.3.1",
-        "rollup-plugin-copy": "^3.4.0",
-        "rollup-plugin-minify-html-literals": "^1.2.6",
-        "rollup-plugin-summary": "^1.3.0",
-        "rollup-plugin-terser": "^7.0.2",
-        "svg-inline-loader": "^0.8.2",
-        "typescript": "^4.3.2",
-        "vite": "^2.3.7"
-    }
-}
