+++ templates/default/.gitignore
@@ -1,8 +1,9 @@
 .idea
 .git
-build
-node_modules
 .DS_Store
-dist/
 .history
+.firebase
+build
+dist
+node_modules
 *.log
