+++ templates/default/.vscode/extensions.json
