+++ templates/default/.vscode/settings.json
