+++ templates/default/public/manifest.json
