+++ templates/default/public/service-worker.js
