+++ templates/default/src/base/app-404.ts
