+++ templates/default/src/base/app-localstorage.ts
