+++ templates/default/src/base/styles/default-theme.ts
