+++ templates/default/src/base/styles/lit-shared-styles.ts
