+++ templates/default/src/xconfig/strings/index.ts
