+++ templates/default/src/xconfig/strings/lang/es.ts
