+++ templates/default/src/xconfig/typings.d.ts
