+++ templates/rapid-prototype/README.md
@@ -1,0 +1,3 @@
+# {{name}}
+
+{{description}}
