+++ templates/rapid-prototype/src/index.ts
